+++ lab4/Khalabarchuk_Project/app/authentication/templates/authentication/login.html
@@ -9,7 +9,7 @@
 
     <div class="main-wrapper">
         <main class="main">
-            <form action="{{ url_for("login") }}" method="post" class="login-form" novalidate>
+            <form action="{{ url_for("auth.login") }}" method="post" class="login-form" novalidate>
                 {{ form.csrf_token }}
                 <div class="form-header">
                     <p class="form-title">
