+++ lab4/Khalabarchuk_Project/app/authentication/templates/authentication/sign_up.html
@@ -9,7 +9,7 @@
 
     <div class="main-wrapper">
         <main class="main w-25">
-            <form action="{{ url_for("sign_up") }}" method="post" class="login-form w-100" novalidate>
+            <form action="{{ url_for("auth.sign_up") }}" method="post" class="login-form w-100" novalidate>
                 {{ reg_form.csrf_token }}
                 <div class="form-header">
                     <p class="form-title">
