+++ lab4/Khalabarchuk_Project/app/common/__init__.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+
+
+common_bp = Blueprint("common", __name__, template_folder="templates/common")
+
+from . import views
