+++ lab4/Khalabarchuk_Project/app/common/templates/common/comments-page.html
@@ -11,7 +11,7 @@
     <div class="main-wrapper">
         <main class="main container w-50">
             <form class="row todo-add-part d-flex flex-row align-items-center w-100"
-                  action="{{ url_for("add_comment") }}"
+                  action="{{ url_for("common.add_comment") }}"
                   METHOD="post"
                   novalidate
             >
