+++ lab4/Khalabarchuk_Project/app/common/templates/common/portfolio-main.html
@@ -20,7 +20,7 @@
                         Blanditiis, magnam!
                     </h3>
                 </div>
-                <a class="projects-button default-button" href="{{ url_for('portfolio_projects') }}">Projects</a>
+                <a class="projects-button default-button" href="{{ url_for('common.portfolio_projects') }}">Projects</a>
             </div>
         </main>
     </div>
