+++ lab4/Khalabarchuk_Project/app/common_logic.py
@@ -1,0 +1,7 @@
+import platform
+
+from flask import render_template, request
+
+
+def base_render(template: str, **context):
+    return render_template(template, about_os=platform.platform(), user_agent_info=request.user_agent.string, **context)
