+++ lab4/Khalabarchuk_Project/app/forms/Comment.py
@@ -1,8 +1,8 @@
-from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField
-from wtforms.validators import DataRequired
-
-
-class CommentForm(FlaskForm):
-    comment = StringField("", validators=[DataRequired(message="This field is required.")])
-    submit = SubmitField("Comment")
+# from flask_wtf import FlaskForm
+# from wtforms import StringField, SubmitField
+# from wtforms.validators import DataRequired
+#
+#
+# class CommentForm(FlaskForm):
+#     comment = StringField("", validators=[DataRequired(message="This field is required.")])
+#     submit = SubmitField("Comment")
