+++ lab4/Khalabarchuk_Project/app/todo/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+todo_bp = Blueprint("todo", __name__, template_folder="templates/todo")
+
+from . import views
