+++ lab4/Khalabarchuk_Project/app/user/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+user_bp = Blueprint("user", __name__, template_folder="templates/user")
+
+from . import views
