+++ lab4/Khalabarchuk_Project/app/user/templates/user/all-users.html
@@ -1,0 +1,35 @@
+{% extends 'portfolio-base.html' %}
+
+{% block links %}
+    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='css/all-users.css') }}">
+{% endblock %}
+
+{% block content %}
+
+    <div class="main-wrapper bg-white-gradient">
+        <main class="main">
+            <h3>All users</h3>
+            <div class="table-wrapper bg-white-gradient-revers">
+                <table class="table table-hover" style="--bs-table-bg:transparent;">
+                    <thead>
+                    <tr>
+                        <th scope="col">email</th>
+                        <th scope="col">username</th>
+                        <th scope="col">avatar_image</th>
+                    </tr>
+                    </thead>
+                    <tbody>
+                    {% for user in users %}
+                        <tr>
+                            <td>{{ user.email }}</td>
+                            <td>{{ user.username }}</td>
+                            <td>{{ user.avatar_image }}</td>
+                        </tr>
+                    {% endfor %}
+                    </tbody>
+                </table>
+            </div>
+        </main>
+    </div>
+
+{% endblock %}
