+++ lab4/Khalabarchuk_Project/run.py
@@ -1,5 +1,4 @@
-from app import app
-from entitys.ToDo import db
+from app import app, db
 
 if __name__ == '__main__':
     app.run(debug=False)
